Moved cad/ to interfaces to prepare for implementing the interface in the blender adapter.
