+++ codetocad/adapters/blender/__init__.py
@@ -1,0 +1,7 @@
+from codetocad.core import *
+
+# from codetocad.adapters.blender.cad.assembly.assembly import Assembly
+# from codetocad.adapters.blender.cad.part.part import Part
+# from codetocad.adapters.blender.cad.sketch.sketch import Sketch
+# from codetocad.adapters.blender.cad.edge.edge import Edge
+# from codetocad.adapters.blender.cad.vertex.vertex import Vertex
