+++ codetocad/adapters/blender/blender_actions/vertex_edge_wire.py
@@ -2,9 +2,9 @@ from typing import List, Tuple, Union
 import bpy
 
 from codetocad.adapters.blender.blender_definitions import BlenderLength
-from codetocad.cad.edge.edge import Edge
-from codetocad.cad.vertex.vertex import Vertex
-from codetocad.cad.wire.wire import Wire
+from codetocad.interfaces.cad.edge.edge import Edge
+from codetocad.interfaces.cad.vertex.vertex import Vertex
+from codetocad.interfaces.cad.wire.wire import Wire
 from codetocad.core.dimensions.point import Point
