--- codetocad/cad/__init__.py
@@ -1,7 +1,0 @@
-from codetocad.core import *
-
-from codetocad.cad.assembly.assembly import Assembly
-from codetocad.cad.part.part import Part
-from codetocad.cad.sketch.sketch import Sketch
-from codetocad.cad.edge.edge import Edge
-from codetocad.cad.vertex.vertex import Vertex
