+++ codetocad/interfaces/cad/wire/wire_add.py
@@ -1,10 +1,10 @@
 from typing import TYPE_CHECKING
-from codetocad.cad.edge.edge import Edge
-from codetocad.cad.vertex.vertex import Vertex
+from codetocad.interfaces.cad.edge.edge import Edge
+from codetocad.interfaces.cad.vertex.vertex import Vertex
 from codetocad.core.dimensions.length import Length, LengthType
 
 if TYPE_CHECKING:
-    from codetocad.cad.wire.wire import Wire
+    from codetocad.interfaces.cad.wire.wire import Wire
 
 
 class WireAdd:
