+++ codetocad/interfaces/cad/wire/wire_constraint.py
@@ -1,12 +1,12 @@
 from dataclasses import dataclass
 from typing import TYPE_CHECKING, TypeAlias
 
-from codetocad.cad.edge.edge import Edge
-from codetocad.cad.vertex.vertex import Vertex
+from codetocad.interfaces.cad.edge.edge import Edge
+from codetocad.interfaces.cad.vertex.vertex import Vertex
 
 
 if TYPE_CHECKING:
-    from codetocad.cad.wire.wire import Wire
+    from codetocad.interfaces.cad.wire.wire import Wire
 
 
 @dataclass
